+++ config/tsuprc/tsup.prod.ts
@@ -17,7 +17,7 @@ export default defineConfig({
 				{ from: "./package.json", to: "./package.json" },
 				{ from: "./.npmrc", to: "./.npmrc" },
 				{ from: "./.npmignore", to: "./.npmignore" },
-				{ from: "./README.npm.md", to: "./README.md" }
+				{ from: "./README.md", to: "./README.md" }
 			]
 		})
 	]