+++ src/functions/index.ts
@@ -1,0 +1,4 @@
+export * from "./myErrorCatcher";
+export * from "./myErrorWrapper";
+export * from "./myErrorHandler";
+export * from "./myError";