+++ src/index.ts
@@ -1,3 +1,17 @@
-export { MyError, myErrorCatcher, myErrorHandler, myErrorWrapper } from "@/utils";
-
-export type { TErrorReturn, TFunctionReturn, TMyError, TDataReturn, TMyErrorList, TMyHandler } from "@/utils/types";
+export * from "@/functions";
+export * from "@/types";
+// export type {
+// 	TMyErrorList,
+// 	IMyError,
+// 	IMyErrorAPI,
+// 	IMyErrorRateLimit,
+// 	IMyErrorValidation,
+// 	TCauseError,
+// 	TDetails,
+// 	TBaseError,
+// 	TBaseErrorExt,
+// 	TValidationError,
+// 	TApiError,
+// 	TApiRateLimit,
+// 	TAllMyErrorTypes
+// } from "@/types";