+++ src/types/internal.ts
@@ -1,4 +1,13 @@
+import type { IMyError, TMyErrorList } from "@/types";
+
+/**
+ * `any` - But only allowed in development.
+ * @internal
+ */
 export type TODO = any; // eslint-disable-line @typescript-eslint/no-explicit-any
+/**
+ * @internal
+ */
 export type Prettify<T> = {
 	[K in keyof T]: NonNullable<T[K]>;
 };
@@ -15,30 +24,23 @@ export type ErrorTypesCatched =
 	| SyntaxErrorConstructor
 	| TypeErrorConstructor;
 
-export type TMyError<T = NonNullable<unknown>> = T & {
-	code?: number | string;
-	hint?: {
-		dev?: string;
-		user?: string;
-	};
-	message?: {
-		dev?: string;
-		user?: string;
-	};
-	name?: string;
-};
-
 export type TDataReturn<T> = [T, false];
-export type TErrorReturn<CustomError = NonNullable<unknown>> = [TMyError<CustomError>, true];
+export type TErrorReturn<CustomError = NonNullable<unknown>> = [CustomError & IMyError, true];
 export type TFunctionReturn<T> = Prettify<TDataReturn<T>> | Prettify<TErrorReturn>;
 
-export type TMyErrorList<CustomError = NonNullable<unknown>> = Record<string, Required<TMyError<CustomError>>>;
+// export type TMyErrorList<CustomError = NonNullable<unknown>> = Record<string, Required<IMyError<CustomError>>>;
+
 export type TMyHandler<
 	CustomError,
 	T extends Record<number | string, { code: string & unknown }> & TMyErrorList<CustomError>
 > = Partial<{
 	[K in T[keyof T]["code"]]: (...arguments_: K[]) => TFunctionReturn<unknown>;
 }>;
-// export type TMyHandler<,T extends TMyErrorList<CustomError>> = Partial<{
+// export type TMyHandler<G extends {},T extends TMyErrorList<G>> = Partial<{
 // 	[K in T[keyof T]["code"]]: (...args: K[]) => TFunctionReturn<unknown>;
 // }>;
+// export type TMyHandler<T extends TMyErrorList> = Partial<{
+// 	[K in T[keyof T]["code"]]: (...args: K[]) => TFunctionReturn<unknown>;
+// }>;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type NoInfer<T> = [T][T extends any ? 0 : never];