+++ src/utils/index.ts
@@ -1,4 +1,1 @@
-export * from "./myErrorCatcher";
-export * from "./myErrorWrapper";
-export * from "./myErrorHandler";
-export * from "./myError";
+export * from "./isPromise";