+++ src/utils/isPromise.ts
@@ -1,0 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/ban-types
+export function isPromise(value: Function): Boolean {
+	return value.constructor.name == "AsyncFuncton";
+}
+export default isPromise;